android

+++ iOS/G3MiOSSDK/Commons/Basic/IG3MBuilder.cpp
@@ -811,7 +811,7 @@ GPUProgramManager* IG3MBuilder::getGPUProgramManager() {
 SceneLighting* IG3MBuilder::getSceneLighting() {
   if (_sceneLighting == NULL) {
     //_sceneLighting = new DefaultSceneLighting();
-    _sceneLighting = new CameraFocusSceneLighting(Color::fromRGBA((float)0.3, (float)0.3, (float)0.3, 1.0),
+    _sceneLighting = new CameraFocusSceneLighting(Color::fromRGBA((float)0.3, (float)0.3, (float)0.3, (float)1.0),
                                                   Color::yellow());
   }
   return _sceneLighting;
